Refactor: Separate caching from SQLETL
Refactor the `CachedSQLETL` class to better separate concerns.

The caching functionality has been extracted from the original
`CachedSQLETL` class into a new class of the same name that
inherits from the now non-caching `SQLETL` base class.

This improves modularity and allows for the use of `SQLETL`
without the overhead of caching if it is not required.